flask_requirements.txt
blinker==1.8.2
click==8.1.7
dill==0.3.8
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
Werkzeug==3.0.3
